Update index.html
Native Instruments has developed a new music format to encapsulate the various components of a musical track in one file so that they can be played individually. This opens the door up for a whole new world of DJing and music production. Our talk will be how we were able to create a web based experience to listen to these stems on the Beatport store (feature launching in August). In addition, myself, Rama Poola, can be the musical guest and DJ the night only using stems. Essentially, it will be a blend of DJing and live production. The talk will be given by Joe Sullivan, the lead developer of the feature, and Robert Petro, the product lead.